+++ client/src/components/related-items-and-outfit/components/ProductCard.jsx
@@ -1,98 +1,7 @@
 import { React, useState, useEffect } from 'react';
-import Star from '../../ratings-and-reviews/Star';
 import StarButton from './StarButton';
 import DeleteButton from './DeleteButton';
-
-function PhotoGallery({ photoArr, setPhotoUrl }) {
-  return (
-    photoArr.map((url) => (
-      <li key={url}>
-        <img
-          className="thumphoto"
-          src={url}
-          onError={(e) => { e.target.src = require('../assets/Image_not_available.jpg'); }}
-          alt="header image"
-          width="40"
-          height="40"
-          onClick={() => {
-            console.log('click');
-            setPhotoUrl(url);
-          }}
-        />
-      </li>
-
-    ))
-  );
-}
-
-function CardDetail({
-  ratings, name, originalPrice, salePrice, category, mainUrl, photoArr, setProductId, id,
-}) {
-  const [show, setShow] = useState(false);
-  const [photoUrl, setPhotoUrl] = useState('');
-  useEffect(() => {
-    setPhotoUrl(mainUrl);
-  }, [mainUrl]);
-  const handleClick = () => {
-    setProductId(id);
-  };
-  const rating = Math.round((ratings) * 10) / 10;
-  const ifSale = salePrice !== null;
-  return (
-    <div>
-      <img
-        className="mainImg"
-        src={photoUrl}
-        onError={(e) => { e.target.src = require('../assets/Image_not_available.jpg'); }}
-        alt="header image"
-        onMouseOver={() => setShow(true)}
-        onMouseOut={() => setShow(false)}
-      />
-      {show && <ul className="list-gallery"><PhotoGallery photoArr={photoArr} setPhotoUrl={setPhotoUrl} /></ul>}
-      <p className="category">
-        {category}
-      </p>
-      <p onClick={() => { handleClick(); }} className="name">
-        {name}
-      </p>
-
-      {ifSale && (
-      <span className="price">
-        <p>
-          <s>
-            $
-            {' '}
-            {originalPrice}
-          </s>
-        </p>
-        <p className="salesPrice">
-          $
-          {' '}
-          {salePrice}
-        </p>
-      </span>
-      )}
-      {!ifSale && (
-      <p className="price">
-        $
-        {' '}
-        {originalPrice}
-      </p>
-      )}
-
-      <span className="ratings">
-        <p>
-          ratings:
-          {' '}
-          {rating}
-        </p>
-        <Star rating={rating} />
-      </span>
-
-    </div>
-
-  );
-}
+import CardDetail from './CardDetail';
 
 function ProductCard({
   id, productId, setProductId, ifRelated, setOutfitArr,
@@ -148,7 +57,7 @@ function ProductCard({
   }, [id]);
 
   return (
-    <div className="card">
+    <div>
       <CardDetail
         ratings={ratings}
         name={name}
