+++ client/src/components/related-items-and-outfit/index.jsx
@@ -1,10 +1,14 @@
-import { React, useState } from 'react';
+import { React, useState, useEffect } from 'react';
 import RelatedList from './components/RelatedList';
 import OutfitList from './components/OutfitList';
 
 function RelatedAndOutfit({ productId, setProductId }) {
   const [relateIndex, setRelateIndex] = useState(0);
   const [outfitIndex, setOutfitIndex] = useState(0);
+  useEffect(() => {
+    setRelateIndex(0);
+    setOutfitIndex(0);
+  }, [productId]);
 
   return (
     <div className="relatedmodule">
@@ -19,10 +23,6 @@ function RelatedAndOutfit({ productId, setProductId }) {
         relateIndex={relateIndex}
         setRelateIndex={setRelateIndex}
       />
-
-      <p> ++++++++</p>
-
-      <p className="sub-title" id="outfitTitle">outfit list</p>
 
       <OutfitList
         className="listAndArrow"
